Day 10 - refactor code restructure
